requirements.txt
streamlit>=1.24
tensorflow-cpu==2.12.0
numpy==1.23.5
Pillow>=9.5
